edol-cli/pyproject.toml
[tool.poetry]
name = "edol-cli"
version = "0.1.0"
description = ""
authors = ["Cristian Dinu <c.dinu@ucl.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.15.1"
toml = "^0.10.2"
pydantic = "^2.10.4"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
edol = "edol_cli.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
